feat: updated to load configuration during runtim
