+++ apps/faucet/.env.sample
@@ -1,9 +1,3 @@
-# Faucet API URL Default
-NAMADA_INTERFACE_FAUCET_API_URL=http://127.0.0.1:5000
-
-# Override the endpoint
-NAMADA_INTERFACE_FAUCET_API_ENDPOINT=/api/v1/faucet
-
 # Specify a proxy port, if proxy is needed (defaults to 9000)
 NAMADA_INTERFACE_PROXY_PORT=1234
