+++ apps/faucet/public/config.toml
@@ -1,0 +1,6 @@
+# Override any of the following with custom values. Defaults are shown:
+
+#base_url = "http://127.0.0.1:5000"
+#endpoint = "/api/v1/faucet"
+#limit = 1_000_000_000
+
