+++ apps/faucet/src/App/Common/AppBanner.tsx
@@ -1,9 +1,9 @@
-import { AppContext } from "App/App";
+import { FaucetAppContext } from "App/FaucetApp";
 import React, { useContext } from "react";
 import { Banner, BannerContents } from "./Banner.components";
 
 export const AppBanner: React.FC = () => {
-  const { isTestnetLive, settings } = useContext(AppContext)!;
+  const { isTestnetLive, settings } = useContext(FaucetAppContext)!;
   return (
     <>
       {!isTestnetLive && settings?.startsAtText && (
