+++ apps/faucet/src/App/SettingsForm.tsx
@@ -1,5 +1,5 @@
 import { ActionButton, Input } from "@namada/components";
-import { AppContext } from "App/App";
+import { FaucetAppContext } from "App/FaucetApp";
 import React, { useContext, useEffect, useState } from "react";
 import { endpoint } from "utils";
 import {
@@ -11,7 +11,7 @@ import {
 
 export const SettingsForm: React.FC = () => {
   const [isFormValid, setIsFormValid] = useState(false);
-  const { setIsModalOpen, baseUrl, setUrl } = useContext(AppContext)!;
+  const { setIsModalOpen, baseUrl, setUrl } = useContext(FaucetAppContext)!;
   const [apiUrl, setApiUrl] = useState(baseUrl);
 
   useEffect(() => {
