+++ apps/faucet/src/App/index.ts
@@ -2,3 +2,4 @@ export { App } from "./App";
 export * from "./App.components";
 export { FaucetForm } from "./Faucet";
 export * from "./Faucet.components";
+export { FaucetApp } from "./FaucetApp";
