+++ apps/faucet/src/config/types.ts
@@ -1,0 +1,6 @@
+export type Config = {
+  base_url: string;
+  endpoint: string;
+  url: string;
+  limit: number;
+};
